+++ web/src/body.rs
@@ -1,6 +1,8 @@
 //! http body types and traits.
 
-pub use xitca_http::body::{Body, BodyExt, BoxBody, Frame, Full, RequestBody, ResponseBody, SizeHint, StreamDataBody};
+pub use xitca_http::body::{
+    Body, BodyExt, BoxBody, Empty, Frame, Full, RequestBody, ResponseBody, SizeHint, StreamDataBody,
+};
 
 pub(crate) use xitca_http::body::Either;