+++ web/src/handler/types/body.rs
@@ -3,15 +3,13 @@
 use core::{cmp, convert::Infallible, pin::pin};
 
 use crate::{
-    body::{BodyExt, BodyStream, BoxBody, ResponseBody},
+    body::{BodyExt, BodyStream, BoxBody, ResponseBody, SizeHint},
     bytes::{Bytes, BytesMut},
     context::WebContext,
     error::{BodyOverFlow, Error},
     handler::{FromRequest, Responder},
     http::{IntoResponse, WebResponse},
 };
-
-use super::header::{self, HeaderRef};
 
 pub struct Body<B>(pub B);
 
@@ -43,29 +41,23 @@ macro_rules! from_bytes_impl {
             type Error = Error;
 
             async fn from_request(ctx: &'a WebContext<'r, C, B>) -> Result<Self, Self::Error> {
-                let limit = HeaderRef::<'a, { header::CONTENT_LENGTH }>::from_request(ctx)
-                    .await
-                    .ok()
-                    .and_then(|header| header.to_str().ok().and_then(|s| s.parse().ok()))
-                    // when content length is 0 the http library should be producing an immediate
-                    // yielding streaming body which result in an empty body collection type.
-                    .map(|len| cmp::min(len, LIMIT))
-                    .unwrap_or_else(|| LIMIT);
-
                 let body = ctx.take_body_ref();
+
+                let limit = match body.size_hint() {
+                    SizeHint::Exact(size) if LIMIT > 0 => cmp::min(size as usize, LIMIT),
+                    SizeHint::Exact(size) => size as usize,
+                    _ => LIMIT,
+                };
 
                 let mut body = pin!(body);
 
-                let mut buf = <$type>::with_capacity(LIMIT);
+                let mut buf = <$type>::with_capacity(limit);
 
-                while let Some(frame) = body.as_mut().frame().await {
-                    let frame = frame.map_err(Into::into)?;
-
-                    if let Ok(data) = frame.into_data() {
-                        buf.extend_from_slice(data.as_ref());
-                        if limit > 0 && buf.len() > limit {
-                            return Err(Error::from(BodyOverFlow { limit }));
-                        }
+                while let Some(data) = body.as_mut().data().await {
+                    let data = data.map_err(Into::into)?;
+                    buf.extend_from_slice(data.as_ref());
+                    if limit > 0 && buf.len() > limit {
+                        return Err(Error::from(BodyOverFlow { limit }));
                     }
                 }